Fixed Jumping and Added Run Animation
Jumping was fixed by creating a new vector with only the flat velocity of the player, and using this as the new velocity while the player was in the air so as to not affect the falling speed.

Animation still needs to be adjusted so that the backwards running animation plays when the player moves back. Currently the forward run plays anytime the player moves.