Pause, Inventory, and Save
I added the ability to pause the game by pressing Esc. This works by setting the time scale to 0 so that the game freezes and unfreezes when you press Esc. again.

I also started to implement a save functionality, but currently it only saves the players money. It shouldn't be too hard to add everything else though

I also OOPified the inventory system. I still have to do some refining, but it should be ready to use and hopefully makes it easier to do stuff with items.